Add title comment at sql script, and removal queries at the end

+++ src/main/resources/create_schema.sql
@@ -1,3 +1,5 @@
+-- Скрипт для создания и настройки базы данных.
+
 CREATE DATABASE IF NOT EXISTS phonebook_database;
 
 CREATE SEQUENCE personal_data_sequence;
@@ -12,7 +14,7 @@ CREATE SEQUENCE phonebook_entry_sequence;
 
 CREATE TABLE phonebook_entry (
     id BIGINT PRIMARY KEY DEFAULT NEXTVAL('phonebook_entry_sequence'),
-    personal_data_id INTEGER UNIQUE NOT NULL REFERENCES personal_data(id)
+    personal_data_id BIGINT UNIQUE NOT NULL REFERENCES personal_data(id)
 );
 
 CREATE SEQUENCE phone_number_sequence;
@@ -21,8 +23,15 @@ CREATE TABLE phone_number (
     id BIGINT PRIMARY KEY DEFAULT NEXTVAL('phone_number_sequence'),
     number CHARACTER VARYING (64) UNIQUE,
     phone_number_type CHARACTER (4),
-    phonebook_entry_id INTEGER NOT NULL REFERENCES phonebook_entry(id)
+    phonebook_entry_id BIGINT NOT NULL REFERENCES phonebook_entry(id)
 );
 
 CREATE USER phonebook_user WITH PASSWORD 'password';
 GRANT ALL PRIVILEGES ON ALL TABLES IN SCHEMA public TO phonebook_user;
+
+-- DROP TABLE phone_number;
+-- DROP SEQUENCE phone_number_sequence;
+-- DROP TABLE phonebook_entry;
+-- DROP SEQUENCE phonebook_entry_sequence;
+-- DROP TABLE personal_data;
+-- DROP SEQUENCE personal_data_sequence;
